remove black hook, spelling fixes, rename LSTM_static module to use lowercase, docstring fixes
