+++ .pre-commit-config.yaml
@@ -39,11 +39,6 @@ repos:
       - id: rst-directive-colons
       - id: rst-inline-touching-normal
       - id: text-unicode-replacement-char
-  - repo: https://github.com/psf/black-pre-commit-mirror
-    rev: 25.1.0
-    hooks:
-      - id: black
-        exclude: ^docs/
   - repo: https://github.com/PyCQA/isort
     rev: 6.0.1
     hooks:
@@ -73,8 +68,6 @@ repos:
     rev: v0.3.9
     hooks:
       - id: blackdoc
-        additional_dependencies: [ 'black==25.1.0' ]
-      - id: blackdoc-autoupdate-black
   - repo: https://github.com/codespell-project/codespell
     rev: v2.4.1
     hooks:
