+++ pyproject.toml
@@ -87,15 +87,6 @@ all = ["xhydro-lstm[dev]", "xhydro-lstm[docs]"]
 
 [tool]
 
-[tool.black]
-target-version = [
-  "py39",
-  "py310",
-  "py311",
-  "py312",
-  "py313"
-]
-
 [tool.bumpversion]
 current_version = "0.1.1-dev.0"
 commit = true
@@ -162,7 +153,7 @@ values = [
 ]
 
 [tool.codespell]
-ignore-words-list = "astroid,socio-economic"
+ignore-words-list = "astroid,socio-economic,statics"
 
 [tool.coverage.paths]
 source = ["src/xhydro_lstm/", "*/site-packages/xhydro_lstm/"]
